Equalize width of <p>, <ul> & <ol> in posts
The wide ul and ol lists look a little out of place next to the
width-limited paragraphs in posts, such as the announcement post for
Berlin at
  /news/2016/12/13/reproducible-builds-berlin-meeting/

Fix this by limiting the lists to the paragraph width as well. It seems
the meeting announcements are the only places where we used bulleted
lists with wide text.