main page: clarify the tagline for what a reproducible build is
As per
https://lists.reproducible-builds.org/pipermail/rb-general/2019-April/001535.html
this seems like it would help prevent confusion, and in some ways (e.g.
we care about all artifacts, not just binaries) it is a lot less
ambiguous.

It would also be nice to point out that yes, the technical definition is
definitely worth a read as there's a lot of complexity to the issue.