Document tar PaxHeader nondeterminism
based on
https://git.savannah.gnu.org/cgit/tar.git/tree/doc/tar.texi?id=f122fc94a7#n10592
https://git.savannah.gnu.org/cgit/tar.git/tree/tests/append02.at?id=f122fc94a7#n64
https://git.savannah.gnu.org/cgit/tar.git/commit/?id=ef0f882382f6

If you create an archive in POSIX format
and the environment variable `POSIXLY_CORRECT` is set, then the
two archives created using the same options on the same set of files
will not be byte-to-byte equivalent even with the above option.  This
is because the posix default for extended header names includes the
PID of the tar process, which is different at each run.