Add continuous tests overview
Previously the tested project where distributed over the /who page, now
there is a single page containing all projects currently tested in the
r-b test infrastructure.

To do so a new field to _data/project.yml called `tests` containing a
link to the r-b test infra.

If this link is available the project will appear on the /citests/
page, ideally replacing the *overview* of t.r-b/debian bottom left.

[ Mariana Moreira ]
 Improve the styling of the page, using bootstrap's cards.

Signed-off-by: Paul Spooren <mail@aparcar.org>
Signed-off-by: Mattia Rizzolo <mattia@debian.org>